% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-signal.R
\name{deprecated-signallers}
\alias{deprecated-signallers}
\alias{signal_stage}
\alias{signal_experimental}
\alias{signal_superseded}
\title{Deprecated functions for signalling lifecycle stages}
\usage{
signal_stage(stage, what, with = NULL, env = caller_env())

signal_experimental(when, what, env = caller_env())

signal_superseded(when, what, env = caller_env())
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
