% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-package.R
\docType{package}
\name{lifecycle-package}
\alias{lifecycle}
\alias{lifecycle-package}
\title{lifecycle: Manage the Life Cycle of your Package Functions}
\description{
Manage the life cycle of your exported functions with shared conventions, documentation badges, and user-friendly deprecation warnings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lifecycle.r-lib.org/}
  \item \url{https://github.com/r-lib/lifecycle}
  \item Report bugs at \url{https://github.com/r-lib/lifecycle/issues}
}

}
\author{
\strong{Maintainer}: Lionel Henry \email{lionel@posit.co}

Authors:
\itemize{
  \item Lionel Henry \email{lionel@posit.co}
  \item Hadley Wickham \email{hadley@posit.co} (\href{https://orcid.org/0000-0003-4757-117X}{ORCID})
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
