% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge}
\alias{badge}
\title{Embed a lifecycle badge in documentation}
\usage{
badge(stage)
}
\arguments{
\item{stage}{A lifecycle stage as a string, one of:
\code{"experimental"}, \code{"maturing"}, \code{"stable"}, \code{"questioning"},
\code{"archived"}, \code{"soft-deprecated"}, \code{"deprecated"}, \code{"defunct"}.}
}
\description{
Use \code{badge()} within a \code{Sexpr} macro to embed a
\href{https://www.tidyverse.org/lifecycle/}{lifecycle} badge in your
documentation. A good place for this badge is at the top of the
topic description:\preformatted{\Sexpr[results=rd, stage=render]{lifecycle::badge("questioning")}
}

The badge is displayed as an image in the HTML version of the
documentation. To make them available in your package, visit
\url{https://github.com/r-lib/rlang/tree/master/man/figures} and copy
all the files starting with \code{lifecycle-} in your \code{man/figures/}
folder.
}
