% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-package.R
\docType{package}
\name{lifecycle-package}
\alias{lifecycle}
\alias{lifecycle-package}
\title{lifecycle: Manage the Life Cycle of your Package Functions}
\description{
Manage the life cycle of your exported functions with
    shared conventions, documentation badges, and non-invasive
    deprecation warnings.
}
\author{
\strong{Maintainer}: Lionel Henry \email{lionel@rstudio.com}

}
\keyword{internal}
