% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-package.R
\docType{package}
\name{lifecycle-package}
\alias{lifecycle}
\alias{lifecycle-package}
\title{lifecycle: Manage the Life Cycle of your Package Functions}
\description{
Manage the life cycle of your exported functions with
    shared conventions, documentation badges, and non-invasive
    deprecation warnings. The 'lifecycle' package defines four
    development stages (experimental, maturing, stable, and
    questioning) and three deprecation stages (soft-deprecated,
    deprecated, and defunct). It makes it easy to insert badges
    corresponding to these stages in your documentation. Usage of
    deprecated functions are signalled with increasing levels of
    non-invasive verbosity.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lifecycle.r-lib.org/}
  \item \url{https://github.com/r-lib/lifecycle}
  \item Report bugs at \url{https://github.com/r-lib/lifecycle/issues}
}

}
\author{
\strong{Maintainer}: Lionel Henry \email{lionel@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
