% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{signal_experimental}
\alias{signal_experimental}
\alias{signal_superseded}
\title{Deprecated funtions for signalling experimental and lifecycle stages}
\usage{
signal_experimental(when, what)

signal_superseded(when, what, with = NULL)
}
\description{
\ifelse{html}{\figure{lifecycle-deprecated.svg}{options: alt='Deprecated lifecycle'}}{\strong{Deprecated}}
Please use \code{\link[=signal_stage]{signal_stage()}} instead
}
\keyword{internal}
