% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{signal_experimental}
\alias{signal_experimental}
\alias{signal_superseded}
\title{Deprecated funtions for signalling experimental and lifecycle stages}
\usage{
signal_experimental(when, what, env = caller_env())

signal_superseded(when, what, env = caller_env())
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Please use \code{\link[=signal_stage]{signal_stage()}} instead
}
\keyword{internal}
