% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{signal_stage}
\alias{signal_stage}
\title{Signal other life cycle changes}
\usage{
signal_stage(stage, what, env = caller_env(2))
}
\arguments{
\item{stage}{Life cycle stage, either "experimental" or "superseded".}

\item{what}{String describing what feature the stage applies too, using
the same syntax as \code{\link[=deprecate_warn]{deprecate_warn()}}.}

\item{env}{Calling environment, used to detect which package to display
in the message. For expert use only.}
}
\description{
\code{signal_stage()} allows you to signal life cycle stages other than
deprecation (for which you should use \code{\link[=deprecate_warn]{deprecate_warn()}} and friends).
There is no behaviour associated with this signal, but in the future
we will provide tools to log and report on usage of experimental and
superseded functions.
}
\examples{
foofy <- function(x, y, z) {
  signal_stage("experimental", foofy())
  x + y / z
}
foofy(1, 2, 3)
}
